+++ src/main/kotlin/de/kaemmelot/datafilesorter/PluginConstants.kt
@@ -1,0 +1,10 @@
+package de.kaemmelot.datafilesorter
+
+object PluginConstants {
+
+    const val PLUGIN_ID = "de.kaemmelot.datafilesorter"
+    const val PLUGIN_PROJECT_SETTINGS_FILE = "DataFileSorter.xml"
+
+    /** File extension for properties files */
+    const val PROPERTIES_EXTENSION = "properties"
+}