+++ src/main/kotlin/de/kaemmelot/datafilesorter/properties/PropertySettings.kt
@@ -1,18 +1,13 @@
 package de.kaemmelot.datafilesorter.properties
 
-import com.intellij.openapi.components.BaseState
+import com.intellij.util.xmlb.annotations.OptionTag
+import de.kaemmelot.datafilesorter.CharConverter
 
-class PropertySettings : BaseState() {
-    var endWithNewline by property(true)
-    var separateGroups by property(true)
-    var indentValues by property(true)
-    var spaceAroundSeparator by property(true)
-    var forceSeparator by property(false)
-    private var forcedSeparatorString by string("=")
-
-    var forcedSeparator: Char
-        get() = forcedSeparatorString?.get(0) ?: '='
-        set(value) {
-            forcedSeparatorString = value.toString()
-        }
-}
+data class PropertySettings(
+    @OptionTag @JvmField val endWithNewline: Boolean = true,
+    @OptionTag @JvmField val separateGroups: Boolean = true,
+    @OptionTag @JvmField val indentValues: Boolean = true,
+    @OptionTag @JvmField val spaceAroundSeparator: Boolean = true,
+    @OptionTag @JvmField val forceSeparator: Boolean = false,
+    @OptionTag(converter = CharConverter::class) @JvmField val forcedSeparator: Char = '='
+)