+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+### Changed
+
+- Removed confirmation for `undo`
+- Used virtual file API to read and write files
+- Persisted state handling now uses `SerializablePersistentStateComponent`
+
 ## 0.1.0 Initial Release
 
 ### Added